Treat timestamps as double rather than integer. Will not multiply timestamp fraction as before. This was a wrong method and cvv values were invalid.

+++ SyncMeasure/Frame.cs
@@ -6,9 +6,9 @@ namespace SyncMeasure
     {
         public int Id { get; }
         public List<Hand> Hands { get; }
-        public int Timestamp { get; }
+        public double Timestamp { get; }
 
-        public Frame(int id, List<Hand> hands, int timestamp)
+        public Frame(int id, List<Hand> hands, double timestamp)
         {
             Id = id;
             Hands = hands ?? new List<Hand>();
